de-duplicate uploadRemote
by creating a new superclass UploaderPlugin
